Move inspector_shared files into inspector_v2
