+++ packages/devtools_app/lib/src/app.dart
@@ -31,8 +31,8 @@ import 'screens/deep_link_validation/deep_links_controller.dart';
 import 'screens/deep_link_validation/deep_links_screen.dart';
 import 'screens/dtd/dtd_tools_controller.dart';
 import 'screens/dtd/dtd_tools_screen.dart';
-import 'screens/inspector_shared/inspector_screen.dart';
-import 'screens/inspector_shared/inspector_screen_controller.dart';
+import 'screens/inspector_v2/inspector_screen.dart';
+import 'screens/inspector_v2/inspector_screen_controller.dart';
 import 'screens/logging/logging_controller.dart';
 import 'screens/logging/logging_screen.dart';
 import 'screens/memory/framework/memory_controller.dart';
