+++ packages/devtools_app/lib/src/screens/inspector_v2/inspector_screen.dart
