+++ packages/devtools_app/lib/src/screens/inspector_v2/inspector_screen_body.dart
@@ -19,9 +19,9 @@ import '../../shared/primitives/blocking_action_mixin.dart';
 import '../../shared/ui/common_widgets.dart';
 import '../../shared/ui/search.dart';
 import '../../shared/utils/utils.dart';
-import '../inspector_shared/inspector_controls.dart';
-import '../inspector_shared/inspector_screen.dart';
 import 'inspector_controller.dart';
+import 'inspector_controls.dart';
+import 'inspector_screen.dart';
 import 'inspector_tree_controller.dart';
 import 'widget_details.dart';
