+++ packages/devtools_app/lib/src/shared/managers/error_badge_manager.dart
@@ -11,7 +11,7 @@ import 'package:devtools_app_shared/utils.dart';
 import 'package:flutter/foundation.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../../screens/inspector_shared/inspector_screen.dart';
+import '../../screens/inspector_v2/inspector_screen.dart';
 import '../../screens/logging/logging_screen.dart';
 import '../../screens/network/network_screen.dart';
 import '../../screens/performance/performance_screen.dart';
